store: rename TxManager trait to HostStoreTxManager
