+++ modules/enclave-api/src/memory.rs
@@ -1,8 +1,7 @@
 use crate::{Enclave, EnclaveCommandAPI, EnclavePrimitiveAPI, EnclaveProtoAPI};
-use store::memory::MemStore;
-use store::transaction::TxManager;
+use store::{host::HostStoreTxManager, memory::MemStore};
 
-impl<'e> TxManager<MemStore> for Enclave<'e, MemStore> {}
+impl<'e> HostStoreTxManager<MemStore> for Enclave<'e, MemStore> {}
 impl<'e> EnclavePrimitiveAPI<MemStore> for Enclave<'e, MemStore> {}
 impl<'e> EnclaveCommandAPI<MemStore> for Enclave<'e, MemStore> {}
 impl<'e> EnclaveProtoAPI<MemStore> for Enclave<'e, MemStore> {}
