+++ modules/enclave-api/src/rocksdb.rs
@@ -1,8 +1,7 @@
 use crate::{Enclave, EnclaveCommandAPI, EnclavePrimitiveAPI, EnclaveProtoAPI};
-use store::rocksdb::RocksDBStore;
-use store::transaction::TxManager;
+use store::{host::HostStoreTxManager, rocksdb::RocksDBStore};
 
-impl<'e> TxManager<RocksDBStore> for Enclave<'e, RocksDBStore> {}
+impl<'e> HostStoreTxManager<RocksDBStore> for Enclave<'e, RocksDBStore> {}
 impl<'e> EnclavePrimitiveAPI<RocksDBStore> for Enclave<'e, RocksDBStore> {}
 impl<'e> EnclaveCommandAPI<RocksDBStore> for Enclave<'e, RocksDBStore> {}
 impl<'e> EnclaveProtoAPI<RocksDBStore> for Enclave<'e, RocksDBStore> {}
