+++ modules/store/src/host.rs
@@ -1,6 +1,6 @@
 use crate::memory::MemStore;
 use crate::prelude::*;
-use crate::transaction::{CommitStore, TxAccessor};
+use crate::transaction::{CommitStore, CreatedTx, TxAccessor, UpdateKey};
 use crate::{KVStore, Result, TxId};
 
 pub enum HostStore {
@@ -55,5 +55,26 @@ impl TxAccessor for HostStore {
             HostStore::RocksDB(store) => store.run_in_mut_tx(tx_id, f),
             HostStore::Memory(store) => store.run_in_mut_tx(tx_id, f),
         }
+    }
+}
+
+/// `HostStoreTxManager` is a transaction manager
+pub trait HostStoreTxManager<S: CommitStore>: HostStoreAccessor<S> {
+    /// `begin_tx` begins ...
+    fn begin_tx(&self, update_key: Option<UpdateKey>) -> Result<<S::Tx as CreatedTx>::PreparedTx> {
+        let tx = self.use_mut_store(|store| store.create_transaction(update_key))?;
+        let tx = tx.prepare()?;
+        self.use_mut_store(|store| store.begin(&tx))?;
+        Ok(tx)
+    }
+
+    /// `commit_tx` commits ...
+    fn commit_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) -> Result<()> {
+        self.use_mut_store(|store| store.commit(tx))
+    }
+
+    /// `rollback_tx` rollback ...
+    fn rollback_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) {
+        self.use_mut_store(|store| store.rollback(tx));
     }
 }
