+++ modules/store/src/transaction.rs
@@ -1,4 +1,3 @@
-use crate::host::HostStoreAccessor;
 use crate::prelude::*;
 use crate::{KVStore, Result, TxId};
 
@@ -60,26 +59,5 @@ pub trait TxAccessor {
     /// `tx_remove` removes key-value pair corresponding to `key` in a specified transaction
     fn tx_remove(&mut self, tx_id: TxId, key: &[u8]) -> Result<()> {
         self.run_in_mut_tx(tx_id, |tx| tx.remove(key))
-    }
-}
-
-/// `TxManager` is a transaction manager
-pub trait TxManager<S: CommitStore>: HostStoreAccessor<S> {
-    /// `begin_tx` begins ...
-    fn begin_tx(&self, update_key: Option<UpdateKey>) -> Result<<S::Tx as CreatedTx>::PreparedTx> {
-        let tx = self.use_mut_store(|store| store.create_transaction(update_key))?;
-        let tx = tx.prepare()?;
-        self.use_mut_store(|store| store.begin(&tx))?;
-        Ok(tx)
-    }
-
-    /// `commit_tx` commits ...
-    fn commit_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) -> Result<()> {
-        self.use_mut_store(|store| store.commit(tx))
-    }
-
-    /// `rollback_tx` rollback ...
-    fn rollback_tx(&self, tx: <S::Tx as CreatedTx>::PreparedTx) {
-        self.use_mut_store(|store| store.rollback(tx));
     }
 }
